Integrasjonstester tilknyttet henlegging flyttet til egen fil
